+++ core/src/main/resources/AppConstants.properties
@@ -391,7 +391,7 @@ stub4testtool.configuration=false
 # setting for changing the default timeout in larva
 larva.timeout=10000
 
-# Larva Test Tool properties to use when deviating from default settings (in TestTool.java)
+# Larva Test Tool properties to use when deviating from default settings (in LarvaTool.java)
 #larva.javasource.directory=${webapp.realpath}/../resources
 #larva.configuration.directory=${webapp.realpath}/../Configuration
 #larva.windiff.command=${webapp.realpath}/../../../../IbisAlgemeenWasbak/WinDiff/WinDiff.Exe
