+++ larva/src/main/java/org/frankframework/larva/HttpServletResponseMock.java
@@ -13,7 +13,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 */
-package org.frankframework.testtool;
+package org.frankframework.larva;
 
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
