+++ larva/src/main/java/org/frankframework/larva/LarvaServlet.java
@@ -13,7 +13,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 */
-package org.frankframework.testtool;
+package org.frankframework.larva;
 
 import java.io.IOException;
 import java.io.InputStream;
