+++ larva/src/main/java/org/frankframework/larva/queues/QueueUtils.java
@@ -13,7 +13,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
  */
-package org.frankframework.testtool.queues;
+package org.frankframework.larva.queues;
 
 import java.lang.reflect.Constructor;
 import java.lang.reflect.Method;
