+++ larva/src/test/java/org/frankframework/larva/TestToolTest.java
@@ -1,4 +1,4 @@
-package org.frankframework.testtool;
+package org.frankframework.larva;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
