+++ test/src/test/testtool/Base64Pipe/scenario01.properties
@@ -2,11 +2,11 @@ scenario.description = txt file
 
 include = common.properties
 
-dir.init.className=nl.nn.adapterframework.testtool.FileSender
+dir.init.className=nl.nn.adapterframework.larva.FileSender
 dir.init.filename= scenario01/init.xml
 dir.init.runAnt=true
 
-file.compare.className=nl.nn.adapterframework.testtool.FileListener
+file.compare.className=nl.nn.adapterframework.larva.FileListener
 file.compare.filename= ${testdata.dir}/b64/file.txt
 file.compare.filename2= ${testdata.dir}/b64/file_out.bin
