+++ test/src/test/testtool/Base64Pipe/scenario02.properties
@@ -2,11 +2,11 @@ scenario.description = pdf file
 
 include = common.properties
 
-dir.init.className=org.frankframework.testtool.FileSender
+dir.init.className=org.frankframework.larva.FileSender
 dir.init.filename= scenario02/init.xml
 dir.init.runAnt=true
 
-file.compare.className=org.frankframework.testtool.FileListener
+file.compare.className=org.frankframework.larva.FileListener
 file.compare.filename= ${testdata.dir}/b64/file.pdf
 file.compare.filename2= ${testdata.dir}/b64/file_out.bin
