+++ test/src/test/testtool/BatchFileTransformerPipe/common.properties
@@ -31,12 +31,12 @@ ignoreContentBetweenKeys5.key2=</fielddefinition>
 removeKey1=&#13;
 removeKey2=&#xD;
 
-fs.insert.className=org.frankframework.testtool.FileSender
+fs.insert.className=org.frankframework.larva.FileSender
 fs.insert.filename=${testdata.dir}/bftp/in/EDIFACT.TXT
 fs.insert.checkDelete=false
 fs.insert.overwrite=true
 
-fs2.insert.className=org.frankframework.testtool.FileSender
+fs2.insert.className=org.frankframework.larva.FileSender
 fs2.insert.filename=${testdata.dir}/bftp/in/CUSTOMER.TXT
 fs2.insert.checkDelete=false
 fs2.insert.overwrite=true
