+++ test/src/test/testtool/Checksum/Adler32/scenario02.properties
@@ -3,7 +3,7 @@ scenario.active=${active.fxf}
 
 include = ../fxf.properties
 
-fxffs.test.className=org.frankframework.testtool.FileSender
+fxffs.test.className=org.frankframework.larva.FileSender
 fxffs.test.filename=${testdata.dir}/Adler32
 fxffs.test.checkDelete=false
 fxffs.test.overwrite=true
