+++ test/src/test/testtool/FilePipe/common.properties
@@ -6,27 +6,27 @@ java.ErrorMessageFormatter.serviceName=ibis4test-ErrorMessageFormatter
 java.FilePipe.className=org.frankframework.senders.IbisJavaSender
 java.FilePipe.serviceName=ibis4test-FilePipe
 
-fs.insert.className=org.frankframework.testtool.FileSender
+fs.insert.className=org.frankframework.larva.FileSender
 fs.insert.filename=${testdata.dir}/dummy.txt
 fs.insert.overwrite=true
 fs.insert.checkDelete=false
 
-fs2.insert.className=org.frankframework.testtool.FileSender
+fs2.insert.className=org.frankframework.larva.FileSender
 fs2.insert.filename=${testdata.dir}/dummy2.txt
 fs2.insert.overwrite=true
 fs2.insert.checkDelete=false
 
-fs3.insert.className=org.frankframework.testtool.FileSender
+fs3.insert.className=org.frankframework.larva.FileSender
 fs3.insert.filename=${testdata.dir}/dummy3.txt
 fs3.insert.overwrite=true
 fs3.insert.checkDelete=false
 
-fs4.insert.className=org.frankframework.testtool.FileSender
+fs4.insert.className=org.frankframework.larva.FileSender
 fs4.insert.filename=${testdata.dir}/xdummy4.txt
 fs4.insert.overwrite=true
 fs4.insert.checkDelete=false
 
-fs5.insert.className=org.frankframework.testtool.FileSender
+fs5.insert.className=org.frankframework.larva.FileSender
 fs5.insert.filename=${testdata.dir}/dummy5.tx1
 fs5.insert.overwrite=true
 fs5.insert.checkDelete=false
