+++ test/src/test/testtool/FileSender/common.properties
@@ -1,6 +1,6 @@
-adapter.antCopyFile.className = org.frankframework.testtool.FileSender
+adapter.antCopyFile.className = org.frankframework.larva.FileSender
 adapter.antCopyFile.filename = scenario01/antScriptThatCopies.xml
 adapter.antCopyFile.runAnt = true
 
-adapter.fileExistenceCheck.className = org.frankframework.testtool.FileListener
+adapter.fileExistenceCheck.className = org.frankframework.larva.FileListener
 adapter.fileExistenceCheck.filename = ${testdata.dir}/FileSender/toCopy.txt
