+++ test/src/test/testtool/FxF3/common.properties
@@ -28,7 +28,7 @@ manage.dbwrapper.serviceName=ibis4test-ManageDatabaseWrapper
 java.delay.className=org.frankframework.senders.IbisJavaSender
 java.delay.serviceName=ibis4test-DelayPipe-5000ms
 
-dir.list.className=org.frankframework.testtool.FileSender
+dir.list.className=org.frankframework.larva.FileSender
 dir.list.filename=${fxf.dir}
 
 ignoreKey1=oracle.sql.BLOB
