+++ test/src/test/testtool/FxF3/scenario12.properties
@@ -3,7 +3,7 @@ scenario.active=${active.jms}
 
 include = common.properties
 
-files.init.className=org.frankframework.testtool.FileSender
+files.init.className=org.frankframework.larva.FileSender
 files.init.filename=12/init.xml
 files.init.runAnt=true
