+++ test/src/test/testtool/ManageFileSystem/scenario01.properties
@@ -1,7 +1,7 @@
 scenario.description = Test all actions - Happy Flow
 include=common.properties
 
-adapter.antCopyFile.className = org.frankframework.testtool.FileSender
+adapter.antCopyFile.className = org.frankframework.larva.FileSender
 adapter.antCopyFile.filename = scenario01/prepare.xml
 adapter.antCopyFile.runAnt = true
