+++ test/src/test/testtool/ManagedFileHandler/common.properties
@@ -27,17 +27,17 @@ manage.db.className=org.frankframework.senders.IbisJavaSender
 manage.db.serviceName=ibis4test-ManageDatabase
 
 # to run ant scripts to move a file etc.
-dir.init.className=org.frankframework.testtool.FileSender
+dir.init.className=org.frankframework.larva.FileSender
 dir.init.runAnt=true
 
-dir.moveFile.className=org.frankframework.testtool.FileSender
+dir.moveFile.className=org.frankframework.larva.FileSender
 dir.moveFile.runAnt=true
 
 # wait sometime for adapter to do its job
 delay.process.className=org.frankframework.senders.IbisJavaSender
 delay.process.serviceName=ibis4test-DelayPipe-5000ms
 
-#check.file.className = org.frankframework.testtool.FileListener
+#check.file.className = org.frankframework.larva.FileListener
 #check.file.filename = ${testdata.dir}/mfh/processed/input.xml-2021-10-06 13_12_40.035
