+++ test/src/test/testtool/RestListener/common.properties
@@ -12,7 +12,7 @@ java.RestSender_GetDoc.serviceName=ibis4test-RestSender_GetDoc
 java.RestSender_GetDoc3.className=org.frankframework.senders.IbisJavaSender
 java.RestSender_GetDoc3.serviceName=ibis4test-RestSender_GetDoc3
 
-dir.list.className=org.frankframework.testtool.FileSender
+dir.list.className=org.frankframework.larva.FileSender
 dir.list.filename=${testdata.dir}/rest
 
 ignoreContentBetweenKeys1.key1 = <rowsupdated>
