+++ test/src/test/testtool/RestListener/scenario01.properties
@@ -7,7 +7,7 @@ ignoreContentBetweenKeys9.key2 = "
 ignoreContentBetweenKeys10.key1=<field name="LENGTH">
 ignoreContentBetweenKeys10.key2=</field>
 
-dir.init.className=org.frankframework.testtool.FileSender
+dir.init.className=org.frankframework.larva.FileSender
 dir.init.filename=01/dir-init.xml
 dir.init.runAnt=true
