+++ test/src/test/testtool/RestListener/scenario02.properties
@@ -2,7 +2,7 @@ scenario.description = OK (streamResultToFileNameSessionKey)
 
 include = common.properties
 
-dir.init.className=org.frankframework.testtool.FileSender
+dir.init.className=org.frankframework.larva.FileSender
 dir.init.filename=02/dir-init.xml
 dir.init.runAnt=true
