+++ test/src/test/testtool/TransactionHandling/scenario03.properties
@@ -3,7 +3,7 @@ scenario.active=${active.fxf}
 
 include = common.properties
 
-fxffs.test.className=org.frankframework.testtool.FileSender
+fxffs.test.className=org.frankframework.larva.FileSender
 fxffs.test.filename=${testdata.dir}/TIMEOUTFXF
 fxffs.test.checkDelete=false
 fxffs.test.overwrite=true
