+++ test/src/test/testtool/UploadFile/common.properties
@@ -3,10 +3,10 @@ include = ../global.properties
 adapter.FileSystem.className=org.frankframework.senders.IbisJavaSender
 adapter.FileSystem.serviceName=${uploadFile-jl.serviceName}
 
-dir.list.className=org.frankframework.testtool.FileSender
+dir.list.className=org.frankframework.larva.FileSender
 dir.list.filename=${testdata.dir}/UploadFile
 
-dir.init.className=org.frankframework.testtool.FileSender
+dir.init.className=org.frankframework.larva.FileSender
 dir.init.filename=dir-init.xml
 dir.init.runAnt=true
