+++ test/src/test/testtool/UploadFile/scenario04.properties
@@ -2,7 +2,7 @@ scenario.description = Clean up the destination folder before uploading file - H
 scenario.active=${active.uploadfile}
 include=common.properties
 
-fs1.create.className = org.frankframework.testtool.FileSender
+fs1.create.className = org.frankframework.larva.FileSender
 fs1.create.filename = ${testdata.dir}/UploadFile/newFile.txt
 fs1.create.overwrite = true
 fs1.create.checkDelete = false
