+++ test/src/test/testtool/UploadFile/scenario05.properties
@@ -2,7 +2,7 @@ scenario.description = Upload a file which already exists in the destination - E
 scenario.active=${active.uploadfile}
 include=common.properties
 
-fs1.create.className = org.frankframework.testtool.FileSender
+fs1.create.className = org.frankframework.larva.FileSender
 fs1.create.filename = ${testdata.dir}/UploadFile/testFile1.txt
 fs1.create.overwrite = true
 fs1.create.checkDelete = false
