+++ test/src/test/testtool/WebServiceListenerSender/scenario11b.properties
@@ -2,7 +2,7 @@ scenario.description = WebServiceListenerSender Address Attachments Multipart Se
 
 include = common.properties
 
-files.init.className=org.frankframework.testtool.FileSender
+files.init.className=org.frankframework.larva.FileSender
 files.init.filename=scenario11b/init.xml
 files.init.runAnt=true
