+++ test/src/test/testtool/WebServiceListenerSender/scenario11d.properties
@@ -2,7 +2,7 @@ scenario.description = WebServiceListenerSender Address Returning Attachments by
 
 include = common.properties
 
-files.init.className=org.frankframework.testtool.FileSender
+files.init.className=org.frankframework.larva.FileSender
 files.init.filename=scenario11d/init.xml
 files.init.runAnt=true
