+++ test/src/test/testtool/WsdlGeneratorPipe/common.properties
@@ -3,7 +3,7 @@ include = ../global.properties
 java.WsdlGeneratorPipe.className=org.frankframework.senders.IbisJavaSender
 java.WsdlGeneratorPipe.serviceName=ibis4test-WsdlGeneratorPipe
 
-dir.list.className=org.frankframework.testtool.FileSender
+dir.list.className=org.frankframework.larva.FileSender
 dir.list.filename=${testdata.dir}/generateWsdl
 
 ignoreContentBetweenKeys1.key1=<wsdl:documentation>
