+++ test/src/test/testtool/WsdlGeneratorPipe/scenario01.properties
@@ -2,7 +2,7 @@ scenario.description = OK (cmhVersion 1)
 
 include = common.properties
 
-dir.init.className=org.frankframework.testtool.FileSender
+dir.init.className=org.frankframework.larva.FileSender
 dir.init.filename=01/dir-init.xml
 dir.init.runAnt=true
