+++ test/src/test/testtool/WsdlGeneratorPipe/scenario02.properties
@@ -2,7 +2,7 @@ scenario.description = OK (cmhVersion 2)
 
 include = common.properties
 
-dir.init.className=org.frankframework.testtool.FileSender
+dir.init.className=org.frankframework.larva.FileSender
 dir.init.filename=02/dir-init.xml
 dir.init.runAnt=true
