+++ test/src/test/testtool/WsdlGeneratorPipe/scenario03.properties
@@ -2,7 +2,7 @@ scenario.description = OK (no ServiceContext)
 
 include = common.properties
 
-dir.init.className=org.frankframework.testtool.FileSender
+dir.init.className=org.frankframework.larva.FileSender
 dir.init.filename=03/dir-init.xml
 dir.init.runAnt=true
