+++ test/src/test/testtool/XmlFileElementIteratorPipe/scenario01.properties
@@ -2,7 +2,7 @@ scenario.description = OK - name
 
 include = common.properties
 
-files.init.className=org.frankframework.testtool.FileSender
+files.init.className=org.frankframework.larva.FileSender
 files.init.filename=scenario01/init.xml
 files.init.runAnt=true
