+++ test/src/test/testtool/XmlFileElementIteratorPipe/scenario02.properties
@@ -2,7 +2,7 @@ scenario.description = OK - chain
 
 include = common.properties
 
-files.init.className=org.frankframework.testtool.FileSender
+files.init.className=org.frankframework.larva.FileSender
 files.init.filename=scenario02/init.xml
 files.init.runAnt=true
