+++ test/src/test/testtool/Zip/ZipIterator/scenario 02.properties
@@ -5,7 +5,7 @@ include = ../common.properties
 provider.java.param1.name = stream
 provider.java.param1.valuefile = simple nested.zip
 
-dir.init.className=org.frankframework.testtool.FileSender
+dir.init.className=org.frankframework.larva.FileSender
 dir.init.filename=../dir-init.xml
 dir.init.runAnt=true
