+++ test/src/test/testtool/Zip/ZipWriter/scenario 01.properties
@@ -2,7 +2,7 @@ scenario.description = ZipWriter, simple zip
 
 include = ../common.properties
 
-dir.init.className=org.frankframework.testtool.FileSender
+dir.init.className=org.frankframework.larva.FileSender
 dir.init.filename=../dir-init.xml
 dir.init.runAnt=true
