+++ test/src/test/testtool/Zip/common.properties
@@ -12,7 +12,7 @@ removeKeysAndContentBetweenKeys3.key2=]
 provider.java.className=org.frankframework.senders.IbisJavaSender
 provider.java.serviceName=ibis4test-unzip
 
-dir.list.className=org.frankframework.testtool.FileSender
+dir.list.className=org.frankframework.larva.FileSender
 dir.list.filename=${testdata.dir}/zip
 
 ignoreContentBetweenKeys1.key1=size="
